requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
fastapi==0.100.1
flower==2.0.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
humanize==4.7.0
idna==3.4
kombu==5.3.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.6
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.4.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
